// src/utils/live-tunnel.d.ts
export declare const startLiveTunnel: ({ localPort, netlifyApiToken, siteId, slug }: {
    localPort: any;
    netlifyApiToken: any;
    siteId: any;
    slug: any;
}) => Promise<any>;
export declare const getLiveTunnelSlug: (state: any, override: any) => any;
//# sourceMappingURL=live-tunnel.d.ts.map